Added Soundcloud option for audio widget

+++ happyweb_system/admin/widgets/audio/delete.php
@@ -1,3 +1,5 @@
 <?php
 // delete the audio file
-unlink($_SERVER["DOCUMENT_ROOT"]."/my_website/uploaded_files/audio/".$data->file);
+if ($data->type == "upload" && $data->file != "") {
+  unlink($_SERVER["DOCUMENT_ROOT"]."/my_website/uploaded_files/audio/".$data->file);
+}
